pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.6.3",
    "httpx>=0.28.1",
    "mcp>=1.1.2",
    "numpy>=2.2.0",
    "pyyaml>=6.0.2",
    "typer>=0.15.1",
]

[project.scripts]
mcp_rag_tool = "rag_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_mcp"]
